split ApiController test
