+++ tests/integration/Core/Framework/Api/Controller/ApiControllerCloneTest.php
@@ -1,0 +1,95 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shopware\Tests\Integration\Core\Framework\Api\Controller;
+
+use Doctrine\DBAL\Connection;
+use PHPUnit\Framework\TestCase;
+use Shopware\Core\Framework\Test\TestCaseBase\AdminApiTestBehaviour;
+use Shopware\Core\Framework\Test\TestCaseBase\DatabaseTransactionBehaviour;
+use Shopware\Core\Framework\Test\TestCaseBase\KernelTestBehaviour;
+use Shopware\Core\Framework\Test\TestCaseHelper\TestUser;
+use Shopware\Core\Framework\Uuid\Uuid;
+use Symfony\Component\HttpFoundation\Response;
+
+/**
+ * @internal
+ */
+class ApiControllerCloneTest extends TestCase
+{
+    use AdminApiTestBehaviour;
+    use DatabaseTransactionBehaviour;
+    use KernelTestBehaviour;
+
+    public function testCloneEntity(): void
+    {
+        $id = Uuid::randomHex();
+        $data = [
+            'id' => $id,
+            'name' => 'test tax clone',
+            'taxRate' => 15,
+        ];
+
+        $this->getBrowser()->jsonRequest('POST', '/api/tax', $data);
+        $response = $this->getBrowser()->getResponse();
+        static::assertSame(Response::HTTP_NO_CONTENT, $response->getStatusCode(), (string) $response->getContent());
+
+        $this->getBrowser()->jsonRequest('GET', '/api/tax/' . $id);
+        $response = $this->getBrowser()->getResponse();
+        static::assertSame(Response::HTTP_OK, $response->getStatusCode(), (string) $response->getContent());
+
+        $tax = json_decode((string) $response->getContent(), true, 512, \JSON_THROW_ON_ERROR);
+        static::assertArrayHasKey('data', $tax);
+        static::assertSame($id, $tax['data']['id']);
+
+        $this->getBrowser()->jsonRequest('POST', '/api/_action/clone/tax/' . $id, $data);
+        $response = $this->getBrowser()->getResponse();
+        static::assertSame(Response::HTTP_OK, $response->getStatusCode(), (string) $response->getContent());
+
+        $data = json_decode((string) $response->getContent(), true, 512, \JSON_THROW_ON_ERROR);
+        static::assertArrayHasKey('id', $data);
+        static::assertNotSame($id, $data['id']);
+
+        $newId = $data['id'];
+        $this->getBrowser()->jsonRequest('GET', '/api/tax/' . $newId);
+        $response = $this->getBrowser()->getResponse();
+        static::assertSame(Response::HTTP_OK, $response->getStatusCode(), (string) $response->getContent());
+
+        $data = json_decode((string) $response->getContent(), true, 512, \JSON_THROW_ON_ERROR);
+        static::assertSame(15.0, $data['data']['attributes']['taxRate']);
+    }
+
+    public function testCloneEntityWithoutPermission(): void
+    {
+        $id = Uuid::randomHex();
+        $data = [
+            'id' => $id,
+            'name' => 'test tax clone',
+            'taxRate' => 15,
+        ];
+
+        $browser = $this->getBrowser();
+        $browser->jsonRequest('POST', '/api/tax', $data);
+        $response = $browser->getResponse();
+        static::assertSame(Response::HTTP_NO_CONTENT, $response->getStatusCode(), (string) $response->getContent());
+
+        $connection = $browser->getContainer()->get(Connection::class);
+        TestUser::createNewTestUser(
+            $connection,
+            ['tax:read']
+        )->authorizeBrowser($browser);
+
+        $browser->jsonRequest('GET', '/api/tax/' . $id);
+        $response = $browser->getResponse();
+        static::assertSame(Response::HTTP_OK, $response->getStatusCode(), (string) $response->getContent());
+
+        $tax = json_decode((string) $response->getContent(), true, 512, \JSON_THROW_ON_ERROR);
+        static::assertArrayHasKey('data', $tax);
+        static::assertSame($id, $tax['data']['id']);
+
+        $browser->jsonRequest('POST', '/api/_action/clone/tax/' . $id, $data);
+        $response = $browser->getResponse();
+        static::assertSame(Response::HTTP_FORBIDDEN, $response->getStatusCode(), (string) $response->getContent());
+    }
+}
