+++ tests/integration/Core/Framework/Api/Controller/ApiControllerVersionTest.php
@@ -1,0 +1,142 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shopware\Tests\Integration\Core\Framework\Api\Controller;
+
+use PHPUnit\Framework\TestCase;
+use Shopware\Core\Content\Product\ProductCollection;
+use Shopware\Core\Content\Product\ProductDefinition;
+use Shopware\Core\Defaults;
+use Shopware\Core\Framework\Api\ApiException;
+use Shopware\Core\Framework\Context;
+use Shopware\Core\Framework\DataAbstractionLayer\EntityRepository;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
+use Shopware\Core\Framework\Test\TestCaseBase\AdminApiTestBehaviour;
+use Shopware\Core\Framework\Test\TestCaseBase\DatabaseTransactionBehaviour;
+use Shopware\Core\Framework\Test\TestCaseBase\KernelTestBehaviour;
+use Shopware\Core\Framework\Uuid\Uuid;
+use Symfony\Component\HttpFoundation\Response;
+
+/**
+ * @internal
+ */
+class ApiControllerVersionTest extends TestCase
+{
+    use AdminApiTestBehaviour;
+    use DatabaseTransactionBehaviour;
+    use KernelTestBehaviour;
+
+    public function testCreateNewVersion(): void
+    {
+        $id = Uuid::randomHex();
+
+        $data = ['id' => $id, 'name' => 'test category'];
+
+        $this->getBrowser()->jsonRequest('POST', '/api/category', $data);
+        $response = $this->getBrowser()->getResponse();
+
+        static::assertSame(Response::HTTP_NO_CONTENT, $response->getStatusCode(), (string) $response->getContent());
+
+        static::assertNotEmpty($response->headers->get('Location'));
+
+        $this->getBrowser()->jsonRequest(
+            'POST',
+            \sprintf('/api/_action/version/category/%s', $id)
+        );
+        $response = $this->getBrowser()->getResponse();
+        $content = json_decode((string) $response->getContent(), true, 512, \JSON_THROW_ON_ERROR);
+
+        static::assertSame(Response::HTTP_OK, $response->getStatusCode(), (string) $response->getContent());
+        static::assertTrue(Uuid::isValid($content['versionId']));
+        static::assertNull($content['versionName']);
+        static::assertSame($id, $content['id']);
+        static::assertSame('category', $content['entity']);
+    }
+
+    public function testDeleteVersion(): void
+    {
+        $id = Uuid::randomHex();
+        $browser = $this->getBrowser();
+
+        $data = [
+            'id' => $id,
+            'productNumber' => Uuid::randomHex(),
+            'stock' => 1,
+            'name' => $id,
+            'tax' => ['name' => 'test', 'taxRate' => 10],
+            'manufacturer' => ['name' => 'test'],
+            'price' => [['currencyId' => Defaults::CURRENCY, 'gross' => 50, 'net' => 25, 'linked' => false]],
+        ];
+
+        $browser->jsonRequest('POST', '/api/product', $data);
+        $response = $browser->getResponse();
+        static::assertSame(Response::HTTP_NO_CONTENT, $browser->getResponse()->getStatusCode(), (string) $browser->getResponse()->getContent());
+        static::assertNotEmpty($response->headers->get('Location'));
+        static::assertSame('http://localhost/api/product/' . $id, $response->headers->get('Location'));
+
+        $this->assertEntityExists($browser, 'product', $id);
+
+        $browser->jsonRequest('POST', '/api/_action/version/product/' . $id);
+        $response = json_decode((string) $browser->getResponse()->getContent(), true, 512, \JSON_THROW_ON_ERROR);
+        static::assertSame(Response::HTTP_OK, $browser->getResponse()->getStatusCode(), (string) $browser->getResponse()->getContent());
+        static::assertIsArray($response);
+        static::assertArrayHasKey('versionId', $response);
+        static::assertArrayHasKey('versionName', $response);
+        static::assertArrayHasKey('id', $response);
+        static::assertArrayHasKey('entity', $response);
+        static::assertTrue(Uuid::isValid($response['versionId']));
+        $versionId = $response['versionId'];
+
+        $browser->jsonRequest('POST', '/api/_action/version/' . $response['versionId'] . '/product/' . $id);
+        $response = json_decode((string) $browser->getResponse()->getContent(), true, 512, \JSON_THROW_ON_ERROR);
+        static::assertSame(Response::HTTP_OK, $browser->getResponse()->getStatusCode(), (string) $browser->getResponse()->getContent());
+        static::assertEmpty($response);
+
+        $this->assertEntityExists($browser, 'product', $id);
+
+        /** @var EntityRepository<ProductCollection> $productRepo */
+        $productRepo = static::getContainer()->get(ProductDefinition::ENTITY_NAME . '.repository');
+        $criteria = new Criteria([$id]);
+        $criteria->addFilter(
+            new EqualsFilter('versionId', $versionId)
+        );
+
+        static::assertCount(0, $productRepo->search($criteria, Context::createDefaultContext()));
+    }
+
+    public function testDeleteVersionWithLiveVersion(): void
+    {
+        $id = Uuid::randomHex();
+        $browser = $this->getBrowser();
+
+        $data = [
+            'id' => $id,
+            'productNumber' => Uuid::randomHex(),
+            'stock' => 1,
+            'name' => $id,
+            'tax' => ['name' => 'test', 'taxRate' => 10],
+            'manufacturer' => ['name' => 'test'],
+            'price' => [['currencyId' => Defaults::CURRENCY, 'gross' => 50, 'net' => 25, 'linked' => false]],
+        ];
+
+        $browser->jsonRequest('POST', '/api/product', $data);
+
+        $browser->jsonRequest('POST', '/api/_action/version/' . Defaults::LIVE_VERSION . '/product/' . $id);
+
+        $repo = static::getContainer()->get(ProductDefinition::ENTITY_NAME . '.repository');
+        $criteria = new Criteria([$id]);
+        $criteria->addFilter(new EqualsFilter('versionId', Defaults::LIVE_VERSION));
+
+        static::assertNotNull($repo->search($criteria, Context::createDefaultContext())->getEntities()->first());
+
+        $response = $browser->getResponse();
+
+        static::assertSame(Response::HTTP_INTERNAL_SERVER_ERROR, $response->getStatusCode(), (string) $response->getContent());
+
+        $content = json_decode((string) $response->getContent(), true, 512, \JSON_THROW_ON_ERROR);
+
+        static::assertSame(ApiException::deleteLiveVersion()->getErrorCode(), $content['errors'][0]['code']);
+    }
+}
